+++ src/main/scala/ot/dispatcher/SuperVisor.scala
@@ -1,14 +1,15 @@
 package ot.dispatcher
 
-import java.sql.ResultSet
-import java.util.Calendar
-
-import org.apache.spark.sql.SparkSession
 import org.apache.log4j.{Level, Logger}
+import org.apache.spark.sql.SparkSession
 import ot.AppConfig
 import ot.AppConfig._
-import ot.dispatcher.sdk.core.CustomException.E00017
+import ot.dispatcher.kafka.context.CommandsContainer
 import ot.dispatcher.sdk.core.CustomException
+import ot.dispatcher.sdk.core.CustomException.E00017
+
+import java.sql.ResultSet
+import java.util.{Calendar, UUID}
 
 /** Gets settings from config file and then runs infinitive loop of user's and system's queries.
  *
@@ -27,29 +28,81 @@ class SuperVisor {
   // Step 1. Loads logger.
   val log: Logger = Logger.getLogger("VisorLogger")
   log.setLevel(Level.toLevel(getLogLevel(config, "visor")))
-  // Step 2. Loads Spark's session and runtime configs.
+  //Step 2. Generate computing node uuid
+  var computingNodeUuid: UUID = getComputingNodeUuid()
+  log.info(s"Computing node uuid: ${computingNodeUuid.toString}")
+  // Step 3. Loads Spark's session and runtime configs.
   val sparkSession: SparkSession = getSparkSession
   log.info("SparkSession started.")
-  // Step 3. Loads connector to DB.
-  val superConnector = new SuperConnector()
-  log.info("SuperConnector is ready.")
-  // Step 4. Loads RAM cache manager.
+  // Step 4. Loads connector to DB.
+  val superDbConnector = new SuperDbConnector()
+  log.info("SuperDbConnector is ready.")
+  //Step 5. Load interactor with Kafka.
+  val kafkaIpAddress: String = config.getString("kafka.ip_address")
+  val kafkaPort: Int = config.getInt("kafka.port")
+  val computingNodeInteractor = new ComputingNodeInteractor(kafkaIpAddress, kafkaPort)
+  //Step 6. Kafka service existing checking
+  val kafkaExists: Boolean = config.getBoolean("kafka.computing_node_mode_enabled")
+  if (kafkaExists) {
+    log.info("SuperKafkaConnector is ready.")
+    log.info("Computing Node Mode is enabled")
+    log.info(s"Kafka ip address: ${kafkaIpAddress}")
+    log.info(s"Kafka port: ${kafkaPort}")
+  } else {
+    log.info("Computing Node Mode is disabled")
+  }
+  // Step 7. Loads RAM cache manager.
   val cacheManager = new CacheManager(sparkSession)
   log.info("CacheManager started.")
-  // Step 5. Loads calculation manager.
-  val superCalculator = new SuperCalculator(cacheManager, superConnector)
+  // Step 8. Loads calculation manager.
+  val superCalculator = new SuperCalculator(cacheManager, superDbConnector)
   log.info("SuperCalculator started.")
 
   /** Starts infinitive loop. */
   def run(): Unit = {
-
     log.info("SuperVisor started.")
     // Step 6. Runs restoring actions after reboot or first start.
     restorationMaintenance()
     log.info("Dispatcher restored DB and RAM Cache.")
-    // Step 7. Runs infinitive loop of system maintenance and user's queries.
+    // Step 7. Register computing node in Kafka
+    if (kafkaExists) {
+      computingNodeInteractor.registerNode(computingNodeUuid)
+      log.info("Spark computing node registered in Kafka")
+    }
+    // Step 8. Runs infinitive loop of system maintenance and user's queries.
     runInfiniteLoop()
+    // Step 9. Unregister computing node in Kafka
+    if (kafkaExists) {
+      computingNodeInteractor.unregisterNode(computingNodeUuid)
+      log.info("Spark computing node unregistered in Kafka")
+    }
+  }
 
+  /**
+   * Get computing node uuid from config or generate it if not exists
+   * @return computing node uuid
+   */
+  def getComputingNodeUuid(): UUID = {
+    var result: UUID = null
+    try {
+      result = UUID.fromString(config.getString("node.uuid"))
+    } catch {
+      case e: Exception => generateNodeUuid()
+    }
+    if (result == null) {
+      generateNodeUuid()
+    } else {
+      result
+    }
+  }
+
+  /**
+   * Generate computing node uuid
+   * @return computing node uuid
+   */
+  private def generateNodeUuid(): UUID = {
+    val bytesContainer: String = "uuid_text"
+    UUID.nameUUIDFromBytes(bytesContainer.getBytes())
   }
 
   /** Returns Spark session and loads config.
@@ -57,7 +110,6 @@ class SuperVisor {
    * @return Spark's session instance.
    */
   def getSparkSession: SparkSession = {
-
     val spark = SparkSession.builder()
       .appName(config.getString("spark.appName"))
       .master(config.getString("spark.master"))
@@ -82,6 +134,11 @@ class SuperVisor {
 
     var negativeDeltaCounter: Int = 0
     val negativeWarnThreshold: Int = config.getInt("loop.negative_warn_threshold")
+
+    new Thread(){
+      override def run(): Unit = {
+      }
+    }
 
     while (true) {
       val delta = Calendar.getInstance().getTimeInMillis - loopEndTime
@@ -109,7 +166,7 @@ class SuperVisor {
   def restorationMaintenance(): Unit = {
     log.trace("Restoration Maintenance section started.")
     val restorationMaintenanceArgs = Map(
-      "superConnector" -> superConnector,
+      "superConnector" -> superDbConnector,
       "cacheManager" -> cacheManager,
       "sparkSession" -> sparkSession
     )
@@ -128,7 +185,7 @@ class SuperVisor {
     log.trace("System Maintenance section started.")
     val systemMaintenanceArgs = Map(
       "cacheManager" -> cacheManager,
-      "superConnector" -> superConnector,
+      "superConnector" -> superDbConnector,
       "sparkSession" -> sparkSession
     )
     val sm = new SystemMaintenance(systemMaintenanceArgs)
@@ -144,19 +201,19 @@ class SuperVisor {
    */
   def userMaintenance(): Unit = {
     // Gets new Jobs.
-    val res = superConnector.getNewQueries
+    val res = superDbConnector.getNewQueries
+    val commandStructs = CommandsContainer.syncValues
 
     import scala.concurrent.{ExecutionContext, Future}
     import ExecutionContext.Implicits.global
-    import scala.util.{Success, Failure}
-
+    import scala.util.{Failure, Success}
     // Starts for each Job calculation process in Future.
     while (res.next()) {
 
       val otlQuery = getOTLQueryObject(res)
       log.info(otlQuery)
       log.debug(s"Job ${otlQuery.id} is setting to running.")
-      superConnector.setJobStateRunning(otlQuery.id)
+      superDbConnector.setJobStateRunning(otlQuery.id)
       val future = Future(futureCalc(otlQuery))
       // Sets logging for future branches depending on it's final state.
       future.onComplete {
@@ -167,7 +224,6 @@ class SuperVisor {
       }
       log.info(s"Job ${otlQuery.id} is running")
     }
-
   }
 
   /** Returns Job ID for logging.
@@ -190,35 +246,35 @@ class SuperVisor {
         superCalculator.calc(otlQuery)
         if (otlQuery.cache_ttl != 0) {
           // Registers new cache id DB.
-          superConnector.addNewCache(otlQuery)
+          superDbConnector.addNewCache(otlQuery)
         }
         // Marks Job in DB as finished.
-        superConnector.setJobStateFinished(otlQuery.id)
+        superDbConnector.setJobStateFinished(otlQuery.id)
         otlQuery.id
 
       } catch {
 
         // Error branch. Marks Job as failed.
         case error: CustomException =>
-          val jobState = superConnector.getJobState(otlQuery.id)
+          val jobState = superDbConnector.getJobState(otlQuery.id)
           if (jobState == "canceled") {
             throw E00017(otlQuery.id)
           } else {
-            superConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
-            superConnector.unlockCaches(otlQuery.id)
+            superDbConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
+            superDbConnector.unlockCaches(otlQuery.id)
             throw error
           }
         case error: OutOfMemoryError =>
-          superConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
-          superConnector.unlockCaches(otlQuery.id)
+          superDbConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
+          superDbConnector.unlockCaches(otlQuery.id)
           throw error
         case error: Exception =>
-          superConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
-          superConnector.unlockCaches(otlQuery.id)
+          superDbConnector.setJobStateFailed(otlQuery.id, error.getLocalizedMessage)
+          superDbConnector.unlockCaches(otlQuery.id)
           throw error
         case throwable: Throwable =>
-          superConnector.setJobStateFailed(otlQuery.id, throwable.getLocalizedMessage)
-          superConnector.unlockCaches(otlQuery.id)
+          superDbConnector.setJobStateFailed(otlQuery.id, throwable.getLocalizedMessage)
+          superDbConnector.unlockCaches(otlQuery.id)
           throw throwable
       }
     }